part2/requirements.txt
flask
flask-restx
pillow==10.4.0